add button component test 2
