+++ src/app/styles/_colors.scss
@@ -2,7 +2,8 @@
   //accent
   --accent-color-100: #73a5ff;
   --accent-color-300: #4c8dff;
-  --accent-color-500: #397df6;
+  //--accent-color-500: #397df6;
+  --color-accent-500: #397df6;
   --accent-color-700: #2f68cc;
   --accent-color-900: #234e99;
