--- src/shared/ui/Button/Button.tsx
@@ -1,5 +1,0 @@
-type ButtonProps = {}
-
-export const Button = (props: ButtonProps) => {
-  return <button>Button</button>
-}
