--- src/shared/ui/stories-example/button.css
@@ -1,38 +1,0 @@
-.storybook-button {
-  cursor: pointer;
-
-  display: inline-block;
-
-  font-family: 'Nunito Sans', 'Helvetica Neue', Helvetica, Arial, sans-serif;
-  font-weight: 700;
-  line-height: 1;
-
-  border: 0;
-  border-radius: 3em;
-}
-
-.storybook-button--primary {
-  color: white;
-  background-color: #1ea7fd;
-}
-
-.storybook-button--secondary {
-  color: #333;
-  background-color: transparent;
-  box-shadow: rgb(0 0 0 / 15%) 0 0 0 1px inset;
-}
-
-.storybook-button--small {
-  padding: 10px 16px;
-  font-size: 12px;
-}
-
-.storybook-button--medium {
-  padding: 11px 20px;
-  font-size: 14px;
-}
-
-.storybook-button--large {
-  padding: 12px 24px;
-  font-size: 16px;
-}
